chore(subnets): always tag the subnets
in the past, we pre-emptively tagged the subnets for use by the eks
cluster. since we didn't have the cluster name in hand, we assumed the
default, `install.id`, would be used as the cluster name. as the new
arch developed, it turned out we did not really _need_ to tag those
subnets ahead of time but it was not disruptive, so we allowed the
pre-emptive tagging to persist.

as it turns out, this pre-emptive tagging _can_ interfere with a number
of components we did not have in-use in house. we are removing that
pre-emptive tagging and therefore must always tag these subnets
regardless of wether or not the cluster name is the default value.

we anticipate this change is not disruptive on the sandbox side.